Add products routes, Add products routers
get list, single and amount,
edit and delete single,
create single

+++ controllers/userController.js
@@ -53,12 +53,14 @@ exports.loginUser = async (req, res) => {
 }
 
 exports.checkIfAdmin = async (req, res) => {
-  res.json({ auth: "admin" })
+  const role = req.userData.role;
+  res.json({ auth: role });
 }
 
 //get list of users for admin panel
 exports.usersList = async (req, res) => {
   try {
+    if(req.userData.role !== "admin") return res.status(400).json({ message: "You have to be an admin" });
     const data = await UserModel.find({}, { pass: 0 })
     res.json(data);
   }
@@ -95,6 +97,7 @@ exports.userInfo = async (req, res) => {
 exports.deleteUser = async (req, res) => {
   const id = req.params.id;
   try {
+    if(req.userData.role !== "admin") return res.status(400).json({ message: "You have to be an admin" });
     if (id != id) {
       const data = await UserModel.deleteOne({ _id: id });
       return res.json(data);
@@ -114,6 +117,7 @@ exports.editUser = async (req, res) => {
   const validBody = validEditUser(req.body);
   if (validBody.error) return res.status(400).json(validBody.error.details);
   try {
+    if(req.userData.role !== "admin") return res.status(400).json({ message: "You have to be an admin" });
     const data = await UserModel.updateOne({ _id: req.params.id }, req.body);
     res.status(201).json(data);
   }
@@ -129,6 +133,7 @@ exports.createUserAsAdmin = async (req, res) => {
     return res.status(400).json(validBody.error.details);
   }
   try {
+    if(req.userData.role !== "admin") return res.status(400).json({ message: "You have to be an admin" });
     const roleName = req.params.role;
     const role = await RoleModel.findOne({roleName});
     if(!role || roleName === "customer"){
